MS_Analytics/requirements.txt
cherrypy==18.8.0
requests==2.28.1
pandas==1.5.3
numpy==1.23.5
statsmodels==0.13.5
python-dotenv==1.0.0
matplotlib==3.7.1
scikit-learn==1.2.2
